requirements.txt
fastapi
uvicorn
motor
pyjwt
python-multipart